requirements/app.txt
galaxy.plugin.api==0.63
requests==2.22.0
psutil==5.6.7
pyobjc-core==5.2; sys_platform == 'darwin'
pyobjc-framework-Cocoa==5.2; sys_platform == 'darwin'
pyobjc-framework-Quartz==5.2; sys_platform == 'darwin'
protobuf==3.7.1

